+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/EnactSpec.hs
@@ -435,7 +435,7 @@ actionPrioritySpec =
 
     -- distinct constitutional values for minFee
     let genMinFeeVals =
-          (\x y z -> (Coin x, Coin y, Coin z))
+          (\x y z -> (CoinPerByte $ CompactCoin x, CoinPerByte $ CompactCoin y, CoinPerByte $ CompactCoin z))
             <$> uniformRM (30, 330)
             <*> uniformRM (330, 660)
             <*> uniformRM (660, 1000)
@@ -450,15 +450,15 @@ actionPrioritySpec =
       pGai0 <-
         submitParameterChange
           SNothing
-          $ def & ppuMinFeeAL .~ SJust val1
+          $ def & ppuTxFeePerByteL .~ SJust val1
       pGai1 <-
         submitParameterChange
           (SJust pGai0)
-          $ def & ppuMinFeeAL .~ SJust val2
+          $ def & ppuTxFeePerByteL .~ SJust val2
       pGai2 <-
         submitParameterChange
           (SJust pGai1)
-          $ def & ppuMinFeeAL .~ SJust val3
+          $ def & ppuTxFeePerByteL .~ SJust val3
       traverse_ @[]
         ( \gaid -> do
             submitYesVote_ (StakePoolVoter spoC) gaid
@@ -469,7 +469,7 @@ actionPrioritySpec =
       getLastEnactedParameterChange
         `shouldReturn` SJust (GovPurposeId pGai2)
       expectNoCurrentProposals
-      getsNES (nesEsL . curPParamsEpochStateL . ppMinFeeAL)
+      getsNES (nesEsL . curPParamsEpochStateL . ppTxFeePerByteL)
         `shouldReturn` val3
 
     it "only the first action of a transaction gets enacted" $ do
@@ -485,15 +485,15 @@ actionPrioritySpec =
           NE.fromList
             [ ParameterChange
                 SNothing
-                (def & ppuMinFeeAL .~ SJust val1)
+                (def & ppuTxFeePerByteL .~ SJust val1)
                 policy
             , ParameterChange
                 SNothing
-                (def & ppuMinFeeAL .~ SJust val2)
+                (def & ppuTxFeePerByteL .~ SJust val2)
                 policy
             , ParameterChange
                 SNothing
-                (def & ppuMinFeeAL .~ SJust val3)
+                (def & ppuTxFeePerByteL .~ SJust val3)
                 policy
             ]
       traverse_
@@ -503,7 +503,7 @@ actionPrioritySpec =
         )
         gaids
       passNEpochs 2
-      getsNES (nesEsL . curPParamsEpochStateL . ppMinFeeAL)
+      getsNES (nesEsL . curPParamsEpochStateL . ppTxFeePerByteL)
         `shouldReturn` val1
       expectNoCurrentProposals