+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -819,8 +819,8 @@ instance
           , sgMaxLovelaceSupply = 45_000_000_000_000_000
           , sgProtocolParams =
               emptyPParams
-                & ppMinFeeAL .~ Coin 44
-                & ppMinFeeBL .~ Coin 155_381
+                & ppTxFeePerByteL .~ CoinPerByte (CompactCoin 44)
+                & ppTxFeeFixedL .~ Coin 155_381
                 & ppMaxBBSizeL .~ 65_536
                 & ppMaxTxSizeL .~ 16_384
                 & ppKeyDepositL .~ Coin 2_000_000