+++ libs/cardano-ledger-api/test/Test/Cardano/Ledger/Api/Tx.hs
@@ -12,6 +12,7 @@ import Cardano.Ledger.Api.Era
 import Cardano.Ledger.Api.PParams
 import Cardano.Ledger.Api.Tx
 import Cardano.Ledger.Binary
+import Cardano.Ledger.Compactible (Compactible (fromCompact))
 import Cardano.Ledger.Core (TxLevel (..))
 import Cardano.Ledger.Hashes (extractHash, hashAnnotated, hashKey)
 import Cardano.Ledger.Keys (makeBootstrapWitness)
@@ -90,12 +91,12 @@ txSpec = describe (eraName @era) $ do
         -- size of the transaction, which in turn affects the overestimation. For this
         -- reason we can only check `>=`
         let
-          overestimatedMinFeeA = toInteger (sum overestimations) <×> pp ^. ppMinFeeAL
+          overestimatedTxFeePerByte = sum overestimations <×> fromCompact (unCoinPerByte (pp ^. ppTxFeePerByteL))
           estimation = estimateMinFeeTx pp tx (Map.size keyPairs) (Map.size byronKeyPairs) 0
           actual = setMinFeeTx pp txSigned 0 ^. bodyTxL . feeTxBodyL
          in
           tabulate "Attrs overestimation in bytes" (map show overestimations) $
-            estimation >= actual <> overestimatedMinFeeA
+            estimation >= actual <> overestimatedTxFeePerByte
 
 spec :: Spec
 spec = do