+++ libs/cardano-ledger-api/test/Test/Cardano/Ledger/Api/Tx/Out.hs
@@ -4,7 +4,6 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeOperators #-}
 
 module Test.Cardano.Ledger.Api.Tx.Out (spec) where
 
@@ -40,6 +39,25 @@ genCompactCoin txOut =
    in
     forAll
       ( CompactCoin
+          <$> oneof [choose (0, 1000000), choose (0, maxCoin), fromIntegral <$> (arbitrary :: Gen Word)]
+      )
+
+genCoinPerByte :: (Testable prop, EraTxOut era) => TxOut era -> (CoinPerByte -> prop) -> Property
+genCoinPerByte txOut =
+  let
+    val = txOut ^. valueTxOutL
+    -- NOTE: As per the UTxO rule:
+    -- ∀ txout ∈ allOuts txb, getValue txout ≥ inject (serSize txout ∗ coinsPerUTxOByte pp)
+    -- Solving for `coinsPerUTxOByte` we get:
+    -- `size (getValue txout) / serSize txout >= coinsPerUTxOByte`
+    -- `serSize txout = 160 + length (serialize protver txout)`, so by substitution:
+    -- `size (getValue txout) / 160 + length (serialize protver txout) >= coinsPerUTxOByte`.
+    -- If we want to maximise this, we can set `length (serialize protver txout)` to `0`, thus:
+    -- `size (getValue txout) / 160 >= coinsPerUTxOByte`
+    maxCoin = quot (fromIntegral $ Val.size val) 160
+   in
+    forAll
+      ( CoinPerByte . CompactCoin
           <$> oneof [choose (0, 1000000), choose (0, maxCoin), fromIntegral <$> (arbitrary :: Gen Word)]
       )
 
@@ -85,12 +103,14 @@ propSetBabbageMinTxOut ::
   ) =>
   Spec
 propSetBabbageMinTxOut =
-  prop "setBabbageMinTxOut" $ \(pp :: PParams era) (txOut :: TxOut era) ->
-    within 1000000 $ -- just in case if there is a problem with termination
-      let txOut' = setMinCoinTxOut pp txOut
-          sz = toInteger (BSL.length (serialize (pvMajor (pp ^. ppProtocolVersionL)) txOut'))
-       in (txOut' ^. coinTxOutL)
-            `shouldBe` Coin ((160 + sz) * unCoin (unCoinPerByte (pp ^. ppCoinsPerUTxOByteL)))
+  prop "setBabbageMinTxOut" $ \(pp' :: PParams era) (txOut :: TxOut era) ->
+    genCoinPerByte txOut $ \cc ->
+      within 1000000 $ -- just in case if there is a problem with termination
+        let pp = pp' & ppCoinsPerUTxOByteL .~ cc
+            txOut' = setMinCoinTxOut pp txOut
+            sz = fromIntegral $ BSL.length (serialize (pvMajor (pp ^. ppProtocolVersionL)) txOut')
+         in (txOut' ^. coinTxOutL)
+              `shouldBe` Coin (fromIntegral $ (160 + sz) * (unCompactCoin . unCoinPerByte) (pp ^. ppCoinsPerUTxOByteL))
 
 propSetEnsureMinTxOutWith ::
   forall era.
@@ -120,6 +140,20 @@ propSetEnsureMinTxOut ::
 propSetEnsureMinTxOut =
   prop "setEnsureMinTxOut" $ propSetEnsureMinTxOutWith @era
 
+propSetBabbageEnsureMinTxOut ::
+  forall era.
+  ( EraTxOut era
+  , Arbitrary (PParamsHKD Identity era)
+  , Arbitrary (TxOut era)
+  , BabbageEraPParams era
+  ) =>
+  Spec
+propSetBabbageEnsureMinTxOut =
+  prop "setBabbageEnsureMinTxOut" $ \(pp :: PParams era) (txOut :: TxOut era) ->
+    genCoinPerByte txOut $ \cc -> do
+      let pp' = pp & ppCoinsPerUTxOByteL .~ cc
+      propSetEnsureMinTxOutWith @era pp' txOut
+
 propSetMaryEnsureMinTxOut :: Spec
 propSetMaryEnsureMinTxOut =
   prop "setMaryEnsureMinTxOut" $ \(txOut :: TxOut MaryEra) ->
@@ -144,10 +178,10 @@ spec =
       propSetEnsureMinTxOut @AlonzoEra
     describe "BabbageEra" $ do
       propSetBabbageMinTxOut @BabbageEra
-      propSetEnsureMinTxOut @BabbageEra
+      propSetBabbageEnsureMinTxOut @BabbageEra
     describe "ConwayEra" $ do
       propSetBabbageMinTxOut @ConwayEra
-      propSetEnsureMinTxOut @ConwayEra
+      propSetBabbageEnsureMinTxOut @ConwayEra
     describe "DijkstraEra" $ do
       propSetBabbageMinTxOut @DijkstraEra
-      propSetEnsureMinTxOut @DijkstraEra
+      propSetBabbageEnsureMinTxOut @DijkstraEra