+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/Gov.hs
@@ -468,8 +468,8 @@ wfPParamsUpdateSpec =
   constrained' $ \ppupdate ->
     -- Note that ppupdate :: SimplePPUpdate
     match ppupdate $
-      \_minFeeA
-       _minFeeB
+      \_txFeePerByte
+       _txFeeFixed
        maxBBSize
        maxTxSize
        maxBHSize