+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Examples/AlonzoAPI.hs
@@ -20,7 +20,7 @@ module Test.Cardano.Ledger.Examples.AlonzoAPI (tests, defaultPParams) where
 import Cardano.Ledger.Alonzo.Tx (alonzoMinFeeTx, hashData)
 import Cardano.Ledger.Alonzo.TxWits (AlonzoEraTxWits (..), TxDats (..), unTxDatsL)
 import Cardano.Ledger.BaseTypes (ProtVer (..), inject)
-import Cardano.Ledger.Coin (Coin (..))
+import Cardano.Ledger.Coin (Coin (..), CompactForm (CompactCoin))
 import Cardano.Ledger.Conway.Core (
   AlonzoEraPParams,
   AlonzoEraTxBody (..),
@@ -35,10 +35,17 @@ import Cardano.Ledger.Conway.Core (
   ppMaxBlockExUnitsL,
   ppMaxTxExUnitsL,
   ppMaxValSizeL,
-  ppMinFeeAL,
+  ppTxFeePerByteL,
   pattern SpendingPurpose,
  )
-import Cardano.Ledger.Core (EraScript (..), EraTx (..), EraTxBody (..), EraTxWits (..), hashScript)
+import Cardano.Ledger.Core (
+  CoinPerByte (..),
+  EraScript (..),
+  EraTx (..),
+  EraTxBody (..),
+  EraTxWits (..),
+  hashScript,
+ )
 import Cardano.Ledger.Hashes (hashAnnotated)
 import Cardano.Ledger.Plutus (ExUnits (..))
 import Cardano.Ledger.Plutus.Data (Data (..))
@@ -90,7 +97,7 @@ testEstimateMinFee =
   where
     pparams =
       defaultPPs emptyPParams
-        & ppMinFeeAL .~ Coin 1
+        & ppTxFeePerByteL .~ CoinPerByte (CompactCoin 1)
     dat = Data (PV1.I 123)
     dataMap = Map.singleton (hashData dat) dat
     script = fromNativeScript $ RequireAllOf mempty