+++ eras/babbage/impl/CHANGELOG.md
@@ -2,12 +2,16 @@
 
 ## 1.13.0.0
 
+* Renamed:
+  - `bppMinFeeA` -> `bppTxFeePerByte`
+  - `bppMinFeeB` -> `bppTxFeeFixed`
+* Changed type of `bppMinFeeA` to `CoinPerByte`
+* Moved `CoinPerByte` to `cardano-ledger-core`
 * Change sets containing errors into `NonEmptySet` for `BabbageUtxowPredFailure`
 * Change all lists into `NonEmpty` for `BabbageUtxoPredFailure`
 * Add `babbageUtxoValidation`
 * Add `babbageUtxoTests`
 * Changed the type of the following fields to `CompactForm Coin` in `BabbagePParams`:
-  - `bppMinFeeA`
   - `bppMinFeeB`
   - `bppKeyDeposit`
   - `bppMinPoolCost`
@@ -44,6 +48,7 @@
 
 ### `testlib`
 
+* Moved `Arbitrary` and `ToExpr` instances of `CoinPerByte` to `cardano-ledger-core`
 * Remove `huddle-cddl` and the `CDDL` modules.
 * Add `plutus_v2_script` to CDDL exports
 * Hide Shelley CDDL `protocol_version` and re-export a new one for Babbage