+++ eras/shelley/impl/CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## 1.18.0.0
 
+* Renamed:
+  - `sppMinFeeA` -> `sppTxFeePerByte`
+  - `ppMinFeeA` -> `ppTxFeePerByte`
+  - `sppMinFeeB` -> `sppTxFeeFixed`
+  - `ppMinFeeB` -> `ppTxFeeFixed`
+* Changed type of `sppMinFeeA` to `CoinPerByte`:
 * Change sets containing errors into `NonEmptySet` for `ShelleyUtxoPredFailure`, `ShelleyUtxowPredFailure`
 * Change all lists into `NonEmpty` for `ShelleyUtxoPredFailure`, `ShelleyUtxowPredFailure`
 * Changed the type of the following fields to `CompactForm Coin` in `ShelleyPParams`:
-  - `sppMinFeeA`
   - `sppMinFeeB`
   - `sppKeyDeposit`
   - `sppMinUTxOValue`
@@ -53,6 +58,9 @@
 
 ### `testlib`
 
+* Renamed:
+  - `maxMinFeeA` -> `maxTxFeePerByte`
+  - `maxMinFeeB` -> `maxTxFeeFixed`
 * Add:
   - `submitBlock_`
   - `submitBlock`